requirements.txt
alabaster==0.7.8
argh==0.26.2
Babel==2.3.3
blinker==1.4
certifi==2016.8.31
cloud-sptheme==1.7.1
colorama==0.3.7
contextlib2==0.5.4
coverage==4.2
Cython==0.24.1
docutils==0.12
dropbox==6.8.0
fake-factory==0.7.2
future==0.15.2
gitdb==0.6.4
github3.py==0.9.3
GitPython==2.0.8
hypothesis==3.4.2
imagesize==0.7.1
Jinja2==2.8
livereload==2.4.1
MarkupSafe==0.23
nose==1.3.7
ntplib==0.3.3
passlib==1.6.5
path.py==8.2.1
pathtools==0.1.2
pefile==2016.3.28
pendulum==0.6.0
port-for==0.3.1
pprp==0.2.5
pycrypto==2.6.1
Pygments==2.1.3
git+https://github.com/132nd-etcher/pyinstaller.git#egg=PyInstaller
pypiwin32==219
PyQt5==5.7
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.12
raven==5.27.0
requests==2.11.1
git+https://github.com/132nd-etcher/ruamel_yaml.git#egg=ruamel.yaml
shellescape==3.4.1
shortuuid==0.4.3
sip==4.18.1
six==1.10.0
smmap==0.9.0
snowballstemmer==1.2.1
sphinx==1.4.1
sphinx-autobuild==0.6.0
sphinx-bootstrap-theme==0.4.12
sphinx-rtd-theme==0.1.9
tornado==4.4.1
typing==3.5.2.2
tzlocal==1.2.2
uritemplate==3.0.0
uritemplate.py==2.0.0
urllib3==1.17
watchdog==0.8.3
WMI==1.4.9